+++ src/Heuristica.hpp
@@ -26,8 +26,8 @@ public:
 	int		NE;
 	ConjuntoConstrucoes ConstrucoesInstancia;
 	int		NV;
-	float	Velocidade;
-	float	TempoDeVidaConcreto;
+	double	Velocidade;
+	double	TempoDeVidaConcreto;
 
 	Heuristica();
 	int		LeDados(string, int );						// le os dados da instancia
@@ -144,7 +144,7 @@ void	Heuristica::ExecutaCONS(string NomeInstancia, int EscolhaVeiculo, int Escol
 	ImprimeSolucao = 0;
 
 	time_t InicioExecucao, FinalExecucao;
-	float TempoExecucao;
+	double TempoExecucao;
 
 	bool ImprimePlanta;
 	bool ImprimeConstrucao;
@@ -341,8 +341,8 @@ void	Heuristica::ExecutaCONS(string NomeInstancia, int EscolhaVeiculo, int Escol
 
 		Solucoes.Solucoes[0].ConstrucoesInstancia.Construcoes[1].Descarregamentos[1].Imprime();
 
-		float HoraInicio;
-		float HoraFinal;
+		double HoraInicio;
+		double HoraFinal;
 		int NumDemanda;
 		int NumPlanta;
 		int Carreta;
@@ -551,7 +551,7 @@ void	Heuristica::ExecutaCONScir(string NomeInstancia, int EscolhaVeiculo, int Es
 	ImprimeSolucao = 0;
 
 	time_t InicioExecucao, FinalExecucao;
-	float TempoExecucao;
+	double TempoExecucao;
 
 	bool ImprimePlanta;
 	bool ImprimeConstrucao;
@@ -592,9 +592,9 @@ void	Heuristica::ExecutaCONScir(string NomeInstancia, int EscolhaVeiculo, int Es
 	//int VerificaViabilidade;
 	//VerificaViabilidade = 1;
 
-	float MakespanAux1;
+	double MakespanAux1;
 	int Buaca1;
-	float MakespanAux2;
+	double MakespanAux2;
 	int Buaca2;
 
 
@@ -753,8 +753,8 @@ void	Heuristica::ExecutaCONScir(string NomeInstancia, int EscolhaVeiculo, int Es
 
 		Solucoes.Solucoes[0].ConstrucoesInstancia.Construcoes[1].Descarregamentos[1].Imprime();
 
-		float HoraInicio;
-		float HoraFinal;
+		double HoraInicio;
+		double HoraFinal;
 		int NumDemanda;
 		int NumPlanta;
 		int Carreta;
@@ -974,7 +974,7 @@ void	Heuristica::ExecutaGrasp(string NomeInstancia, long int NumeroIteracoes, lo
 	ImprimeSolucao = 0;
 
 	time_t InicioExecucao, FinalExecucao;
-	float TempoExecucao;
+	double TempoExecucao;
 
 	//bool ImprimePlanta;
 	//bool ImprimeConstrucao;
@@ -1229,7 +1229,7 @@ void	Heuristica::ExecutaGraspCir(string NomeInstancia, long int NumeroIteracoes,
 	ImprimeSolucao = 0;
 
 	time_t InicioExecucao, FinalExecucao;
-	float TempoExecucao;
+	double TempoExecucao;
 
 	//bool ImprimePlanta;
 	//bool ImprimeConstrucao;
@@ -1274,9 +1274,9 @@ void	Heuristica::ExecutaGraspCir(string NomeInstancia, long int NumeroIteracoes,
 	//int VerificaViabilidade;
 	//VerificaViabilidade = 1;
 
-	float MakespanAux1;
+	double MakespanAux1;
 	int Buaca1;
-	float MakespanAux2;
+	double MakespanAux2;
 	int Buaca2;
 
 	int iteracoes;
@@ -1507,7 +1507,7 @@ void	Heuristica::ExecutaGraspClass(string NomeInstancia, long int NumeroIteracoe
 	ImprimeSolucao = 0;
 
 	time_t InicioExecucao, FinalExecucao;
-	float TempoExecucao;
+	double TempoExecucao;
 
 	//bool ImprimePlanta;
 	//bool ImprimeConstrucao;
@@ -1766,7 +1766,7 @@ void	Heuristica::ExecutaGraspClassCir(string NomeInstancia, long int NumeroItera
 	ImprimeSolucao = 0;
 
 	time_t InicioExecucao, FinalExecucao;
-	float TempoExecucao;
+	double TempoExecucao;
 
 	//bool ImprimePlanta;
 	//bool ImprimeConstrucao;
@@ -1811,9 +1811,9 @@ void	Heuristica::ExecutaGraspClassCir(string NomeInstancia, long int NumeroItera
 	//int VerificaViabilidade;
 	//VerificaViabilidade = 1;
 
-	float MakespanAux1;
+	double MakespanAux1;
 	int Buaca1;
-	float MakespanAux2;
+	double MakespanAux2;
 	int Buaca2;
 
 	int iteracoes;
