+++ src/Procedimento1.hpp
@@ -19,14 +19,14 @@ public:
 	int 	NE;
 	ConjuntoConstrucoes ConstrucoesInstancia;
 	int 	NV;
-	float 	Velocidade;
-	float 	TempoDeVidaConcreto;
+	double 	Velocidade;
+	double 	TempoDeVidaConcreto;
 
 	DadosModelo DM;
 
 	Procedimento1();
 
-	void 	CarregaDados(int InstNP, ConjuntoPlantas InstPlantasInstancia, int InstNE, ConjuntoConstrucoes InstConstrucoesInstancia, int InstNV, float InstVelocidade, float InstTempoDeVidaConcreto, DadosModelo dm);		// carrega os dados da solução
+	void 	CarregaDados(int InstNP, ConjuntoPlantas InstPlantasInstancia, int InstNE, ConjuntoConstrucoes InstConstrucoesInstancia, int InstNV, double InstVelocidade, double InstTempoDeVidaConcreto, DadosModelo dm);		// carrega os dados da solução
 	int 	SelecionaConstrucao(  int &ConstrucaoVaiSerSuprida, int &IndiceConstrucaoVaiSerSuprida);											// Seleciona uma construção baseada em um rank que elas possuem
 	void 	ConfereSeNaoEncontrouUmaPlanta( int  PlantaSelecionada);											// Verifica se não encontrou uma planta
 
@@ -45,7 +45,7 @@ Procedimento1::Procedimento1(){
 }
 
 // carrega os dados da solução
-void 	Procedimento1::CarregaDados(int InstNP, ConjuntoPlantas InstPlantasInstancia, int InstNE, ConjuntoConstrucoes InstConstrucoesInstancia, int InstNV, float InstVelocidade, float InstTempoDeVidaConcreto, DadosModelo dm){
+void 	Procedimento1::CarregaDados(int InstNP, ConjuntoPlantas InstPlantasInstancia, int InstNE, ConjuntoConstrucoes InstConstrucoesInstancia, int InstNV, double InstVelocidade, double InstTempoDeVidaConcreto, DadosModelo dm){
 	NP 	= InstNP;
 	PlantasInstancia 		= InstPlantasInstancia;
 	NE 	= InstNE;
